proxy_service/requirements.txt
redis
grpcio
grpcio-tools